refactor: change structure of Menu folder, add MenuContainer and styled.js
