+++ frontend/src/components/TextField/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './TextField';
