+++ frontend/src/components/layout/Menu/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Menu';
+export { default } from './MenuContainer';
