+++ frontend/src/components/layout/styled.ts
@@ -2,7 +2,7 @@ import styled from 'styled-components';
 import LogoutIcon from '@mui/icons-material/Logout';
 import Grid from '@mui/material/Grid';
 
-const HEADER_HEIGHT = '70px';
+export const HEADER_HEIGHT = '70px';
 
 export const LayoutHeader = styled(Grid)({
   width: 'calc(100vw-10px)',
@@ -10,6 +10,7 @@ export const LayoutHeader = styled(Grid)({
   borderBottom: '2px solid #f0f2f7',
   fontFamily: '"Lato", sans-serif',
 });
+
 export const HeaderDivider = styled('div')({
   display: 'inline-block',
   width: '1px',
@@ -18,6 +19,7 @@ export const HeaderDivider = styled('div')({
   marginTop: '23px',
   marginBottom: '2px',
 });
+
 export const BrandLogo = styled('div')({
   display: 'block',
   height: '68px',
@@ -26,6 +28,7 @@ export const BrandLogo = styled('div')({
   borderRight: '2px solid #f0f2f7',
   color: '#d83938',
 });
+
 export const HeaderContent = styled('div')({
   display: 'grid',
   gridTemplateColumns: '1fr 130px 1px 340px 1px 100px',
@@ -33,9 +36,11 @@ export const HeaderContent = styled('div')({
   maxWidth: '1200px',
   textAlign: 'right',
 });
+
 export const SpaceHolder = styled('div')({
   display: 'inline-block',
 });
+
 export const Counter = styled('div')({
   display: 'inline-block',
   width: '70px',
@@ -48,9 +53,11 @@ export const Counter = styled('div')({
   fontSize: '18px',
   fontFamily: '"Lato", sans-serif',
 });
+
 export const MyCoursesCounterContent = styled('div')({
   display: 'flex',
 });
+
 export const MyCoursesCounterNumber = styled('span')({
   display: 'inline-block',
   height: '18px',
@@ -58,6 +65,7 @@ export const MyCoursesCounterNumber = styled('span')({
   padding: '2px 8px 8px 5px',
   textAlign: 'center',
 });
+
 export const LogOut = styled(LogoutIcon)({
   display: 'inline-block',
   width: '50px',
@@ -68,31 +76,4 @@ export const LogOut = styled(LogoutIcon)({
   '&:hover': {
     cursor: 'pointer',
   },
-});
-export const MenuTabs = styled('div')({
-  width: '100%',
-  height: `calc(100vh - ${HEADER_HEIGHT})`,
-  backgroundColor: 'white',
-  borderRight: '2px solid #f0f2f7',
-  fontSize: '24px',
-  fontFamily: '"Lato", sans-serif',
-  margin: '0px',
-  paddingTop: '20px',
-});
-export const MenuTab = styled('div')({
-  display: 'flex',
-  height: '60px',
-  padding: '14px 20px 18px 20px',
-  borderLeft: '3px solid white',
-  color: 'black',
-  '&:hover': {
-    cursor: 'pointer',
-    color: '#d83938',
-    borderLeft: '3px solid #d83938',
-  },
-});
-export const MenuTabName = styled('span')({
-  display: 'inline-block',
-  height: '24px',
-  padding: '4px 0px 0px 8px',
 });
